requirements.txt
pytest==7.1.3
requests==2.28.1
allure-pytest==2.10.0
selenium==4.1.3
pendulum==2.1.2
pytest-xdist==2.5.0